# NF/TST: add `repo(init=bool)`, and assorted test modernization